Merged all Service handling stuff into Server class, making Service classes all Criteria only.
I like it this way. The interaction between Service and Server was
really ugly and unclear. With Server/Criteria, it's a lot cleaner. It's
much cleaner, and the Criteria API is a lot cleaner as well.

This made it possible to actually support different endpoints as well,
and in a simple way.